+++ release.sh
@@ -58,9 +58,9 @@ function make_release() {
 function install_release() {
     rm -r ~/Zomboid/Workshop/"${MOD_NAME}"
 
-    cp -r  "release/${RELEASE_NAME}" ~/Zomboid/Workshop/"${MOD_NAME}"
+    cp -r "release/${RELEASE_NAME}" ~/Zomboid/Workshop/"${MOD_NAME}"
 
-    rm -r  "release/${RELEASE_NAME}"
+    rm -r "release/${RELEASE_NAME}"
 }
 
 make_release && install_release